Move movement to a separate crate
