+++ collisions/src/prelude.rs
@@ -1,3 +1,4 @@
+pub use super::{ColliderId, CollisionResult, CollisionResultRef};
 pub use crate::colliders::{Collider, ColliderData};
 pub use crate::plugin::{ColliderBundle, ColliderComponent, ColliderDespawnable};
 pub use crate::shapes::{ColliderShape, ColliderShapeType};
