+++ movement/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+"common" = { path = "../common" }
+"collisions" = { path = "../collisions" }
+"bevy" = { version = "0.14"}
