+++ movement/src/paths.rs
@@ -1,11 +1,11 @@
-#![allow(dead_code)]
-
 use std::collections::VecDeque;
 
 use bevy::prelude::*;
 use common::Position;
 
-use super::{steering::SteeringTarget, SteeringHost};
+use crate::plugin::SteeringHost;
+
+use super::SteeringTarget;
 
 #[derive(Debug, Clone, Copy, Reflect)]
 pub struct SteerPathNode {
