+++ movement/src/prelude.rs
@@ -1,0 +1,2 @@
+pub use super::SteeringTarget;
+pub use crate::plugin::{PhysicsParams, SteeringBundle, SteeringHost, SteeringPlugin};
